moved professor page contents to new folder profilePage
