+++ website/package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "antd": "^5.11.2",
+    "antd": "^5.11.4",
     "bootstrap": "^5.3.2",
     "next": "13.5.6",
     "react": "^18",
